cloudfunctions/instance/requirements.txt
google-cloud-bigquery==3.21.0
requests==2.32.3
